+++ CHANGELOG.md
@@ -78,6 +78,7 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 - [#112](https://github.com/ARGOeu/argo-mon-status-api/pull/112) ARGO-5461: Sync Web API Tenants to Local DB (Dev Environment Only)
 - [#115](https://github.com/ARGOeu/argo-mon-status-api/pull/115) ARGO-5477: Support force parameter when creating new topology items
 - [#118](https://github.com/ARGOeu/argo-mon-status-api/pull/118) ARGO-5491 Status-api generate an INIT_CONNECTOR event
+- [#120](https://github.com/ARGOeu/argo-mon-status-api/pull/120) ARGO-5496: Add option for status page theming configuration
 
 ### Fix